Trying to prevent crash seen on PlayStore probably due to friend.getPhoneNumersWithLabel() being called in IO dispatched coroutine
