+++ app/src/main/java/org/linphone/utils/LinphoneUtils.kt
@@ -172,16 +172,5 @@ class LinphoneUtils {
             remoteSipUri.clean()
             return "${localSipUri.asStringUriOnly()}~${remoteSipUri.asStringUriOnly()}"
         }
-
-        fun arePhoneNumberWeakEqual(number1: String, number2: String): Boolean {
-            return trimPhoneNumber(number1) == trimPhoneNumber(number2)
-        }
-
-        private fun trimPhoneNumber(phoneNumber: String): String {
-            return phoneNumber.replace(" ", "")
-                .replace("-", "")
-                .replace("(", "")
-                .replace(")", "")
-        }
     }
 }
