+++ app/src/main/java/org/linphone/utils/PhoneNumberUtils.kt
@@ -127,5 +127,16 @@ class PhoneNumberUtils {
             }
             return null
         }
+
+        fun arePhoneNumberWeakEqual(number1: String, number2: String): Boolean {
+            return trimPhoneNumber(number1) == trimPhoneNumber(number2)
+        }
+
+        private fun trimPhoneNumber(phoneNumber: String): String {
+            return phoneNumber.replace(" ", "")
+                .replace("-", "")
+                .replace("(", "")
+                .replace(")", "")
+        }
     }
 }
